[Runtime] Fix databricks runtime tests and structure (#4036)
