+++ mlrun/runtimes/__init__.py
@@ -31,8 +31,9 @@ from mlrun.runtimes.utils import resolve_spark_operator_version
 from .base import BaseRuntime, RunError, RuntimeClassMode  # noqa
 from .constants import MPIJobCRDVersions
 from .daskjob import DaskCluster  # noqa
+from .databricks.databricks import DatabricksRuntime
 from .function import RemoteRuntime
-from .kubejob import DatabricksRuntime, KubejobRuntime  # noqa
+from .kubejob import KubejobRuntime  # noqa
 from .local import HandlerRuntime, LocalRuntime  # noqa
 from .mpijob import MpiRuntimeContainer, MpiRuntimeV1, MpiRuntimeV1Alpha1  # noqa
 from .nuclio import nuclio_init_hook
